--- a/.markdownlint.json
+++ b/.markdownlint.json
@@ -1,0 +1,5 @@
+{
+  "MD024": {
+    "siblings_only": true
+  }
+}