--- a/jest.config.js
+++ b/jest.config.js
@@ -1,20 +1,23 @@
 /**
  * @fileoverview Jest testing framework configuration
  *
- * Configures Jest for TypeScript testing with:
- * - TypeScript compilation via ts-jest
+ * Configures Jest for ESM + TypeScript testing with:
+ * - TypeScript compilation via ts-jest with ESM support
  * - Module path mapping for clean imports
  * - Test coverage reporting
  * - Test file discovery patterns
  * - Global test setup integration
  */
 
-module.exports = {
-  // Use ts-jest preset for TypeScript support
-  preset: 'ts-jest',
+export default {
+  // Use ts-jest preset with ESM support
+  preset: 'ts-jest/presets/default-esm',
 
   // Node.js environment for testing (not browser)
   testEnvironment: 'node',
+
+  // ESM support configuration
+  extensionsToTreatAsEsm: ['.ts'],
 
   // Root directories to search for source and test files
   roots: ['<rootDir>/src', '<rootDir>/tests'],
@@ -40,14 +43,25 @@ module.exports = {
   // Pattern to match test files
   testMatch: ['**/tests/**/*.test.ts'],
 
-  // Transform TypeScript files using ts-jest
+  // Transform TypeScript files using ts-jest with ESM
   transform: {
-    '^.+\\.ts$': 'ts-jest',
+    '^.+\\.ts$': ['ts-jest', {
+      useESM: true
+    }],
+    '^.+\\.js$': ['ts-jest', {
+      useESM: true
+    }],
   },
+
+  // Transform ESM modules that don't have CommonJS builds
+  transformIgnorePatterns: [
+    'node_modules/(?!.*)'
+  ],
 
   /**
    * Module name mapping for clean imports
    * Maps @-prefixed aliases to actual source paths
+   * ESM requires mapping to actual .ts files during test time
    */
   moduleNameMapper: {
     '^@core/(.*)$': '<rootDir>/src/core/$1',
@@ -59,6 +73,10 @@ module.exports = {
     '^@lib$': '<rootDir>/src/lib/index.ts',
     '^@constants$': '<rootDir>/src/constants/index.ts',
     '^@errors$': '<rootDir>/src/errors/index.ts',
+    '^@utils/(.*)$': '<rootDir>/src/utils/$1',
+    '^@utils$': '<rootDir>/src/utils/index.ts',
+    '^@plugins/(.*)$': '<rootDir>/src/plugins/$1',
+    '^@plugins$': '<rootDir>/src/plugins/index.ts',
   },
 
   /**