## Summary

This PR implements two critical migrations that fundamentally improve the codebase's maintainability and performance:

1. **Migration from marked to remark ecosystem** - Enables AST-based markdown processing
2. **Migration from Jest to Vitest** - Provides native ESM support and significant performance improvements

## Primary Motivation: AST-Based Processing

### The Problem with marked + regex approach

Previously, we used `marked` for markdown parsing, but it had a critical limitation: **it didn't provide access to the markdown AST (Abstract Syntax Tree)**. This forced us to implement document modifications using regex pattern matching, which was:

- **Error-prone**: Regex patterns could match unintended content
- **Fragile**: Complex markdown structures broke our patterns
- **Limited**: Difficult to implement sophisticated transformations
- **Unmaintainable**: Hard to debug and extend functionality

### The Solution: remark ecosystem

The migration to `remark` (part of the unified ecosystem) provides:

- **Full AST access**: Complete programmatic access to markdown structure
- **Robust transformations**: Precise node-based modifications instead of string manipulation
- **Custom plugins**: Ability to create purpose-built plugins for legal document processing
- **Composable architecture**: Chain transformations safely and predictably

### Example: Before vs After

**Before (marked + regex):**
```javascript
// Error-prone regex pattern matching
content = content.replace(/{{#if\s+([^}]+)}}(.*?){{\/if}}/g, (match, condition, content) => {
  // Hope the regex doesn't match nested conditions or break on edge cases
  return evaluateCondition(condition) ? content : '';
});
```

**After (remark + AST):**
```javascript
// Precise AST-based transformation
export function remarkClauses(options) {
  return (tree) => {
    visit(tree, 'text', (node, index, parent) => {
      // Safely process only text nodes containing conditionals
      // Full context awareness of document structure
      processConditionalBlocks(node, options.metadata);
    });
  };
}
```

## Key Technical Improvements

### Remark Migration Benefits

- **AST-based processing**: Eliminates regex-based document manipulation
- **Plugin architecture**: Modular, testable, and extensible transformations
- **Type safety**: Full TypeScript support with proper AST node types
- **ESM compatibility**: Native ES modules support
- **Future-proof**: Active ecosystem with regular updates

### Vitest Migration Benefits

- **Native ESM support**: No complex transformation workarounds needed
- **Performance**: 3-5x faster test execution (40s → 15-20s full suite)
- **Modern tooling**: Better debugging, watch mode, and developer experience
- **Simplified configuration**: Removed complex Jest ESM compatibility layers

## Technical Changes

### Core Architecture
- Replaced `marked` with `unified` + `remark-parse` + `remark-stringify`
- Implemented AST-based plugins for legal document processing:
  - `remarkClauses`: Conditional content processing
  - `remarkVariables`: Variable substitution
  - `remarkStructure`: Document structure management

### Test Infrastructure
- Migrated from Jest to Vitest for native ESM compatibility
- Updated all test files from `jest.*` to `vi.*` syntax
- Simplified test configuration (removed ESM transformation workarounds)
- Enhanced test timeouts for CI environments

### Documentation
- Comprehensive migration guides and architecture documentation
- Updated development guidelines with new AST-based approach
- Enhanced plugin development documentation

## Performance Metrics

| Metric | Before | After | Improvement |
|--------|--------|-------|-------------|
| Full test suite | ~40+ seconds | ~15-20 seconds | **3x faster** |
| Unit tests | ~15-20 seconds | ~2-5 seconds | **5x faster** |
| Document processing | Regex-based (fragile) | AST-based (robust) | **Reliability** |
| ESM Support | Complex workarounds | Native support | **Simplified** |

## Test Results

- ✅ All 1295 tests passing
- ✅ TypeScript compilation successful  
- ✅ Pre-commit hooks validated
- ✅ ESM compatibility verified
- ✅ AST-based transformations working correctly

## Breaking Changes

- Minimum Node.js version may be affected due to ESM requirements
- Plugin API changed from regex-based to AST-based (internal changes only)
- Test command syntax updated (npm scripts remain the same)

## Migration Impact

This migration positions the codebase for:
- More reliable document processing
- Easier implementation of complex legal document features
- Better maintainability and debugging capabilities
- Future enhancements based on document structure analysis