@diegomarino 2025-07-31
:tada: This PR is included in version 2.15.0 :tada:

The release is available on:
- [npm package (@latest dist-tag)](https://www.npmjs.com/package/legal-markdown-js/v/2.15.0)
- [GitHub release](https://github.com/petalo/legal-markdown-js/releases/tag/v2.15.0)

Your **[semantic-release](https://github.com/semantic-release/semantic-release)** bot :package::rocket: